+++ Minecraft/src/Mob.h
@@ -9,6 +9,6 @@ public:
 
 	Mob(glm::vec3 position, int health, int maxHealth);
 	
-	virtual void Update(float deltaTime) = 0;
+	virtual void Update(double deltaTime) = 0;
 	virtual void Render(Shader* shader) = 0;
 };
